// src/core-rpg/types/combat-action.type.ts
export enum CombatAction {
    // Basic actions
    QuickAssist = 'quick-assist',
    ExSpecialAttack = 'ex-special-attack',
    ChainAttack = 'chain-attack',
    // Anomalies
    Freeze = 'freeze',
    Shatter = 'shatter',
    Assault = 'assault',
    AttackOnCorruption = 'attack-on-corruption',
    AttackOnBurning = 'attack-on-burning',
    // Rotations
    BeingOffField = 'being-off-field',
    SwitchOnField = 'switch-on-field',
    EnteringCombat = 'entering-combat',
}