// src/core-rpg/types/combat-stat.type.ts
export enum CombatStat {
    Health = 'health',
    Attack = 'attack',
    Defense = 'defense',
    Damage = 'damage',
    EtherDamage = 'ether-damage',
    FireDamage = 'fire-damage',
    IceDamage = 'ice-damage',
    ElectricDamage = 'electric-damage',
    PhysicalDamage = 'physical-damage',
    CriticalDamage = 'critical-damage',
    CriticalRate = 'critical-rate',
    AnomalyMastery = 'anomaly-mastery',
    AnomalyProficiency = 'anomaly-proficiency',
    Daze = 'daze',
    Impact = 'impact',
    Shield = 'shield',
    Penetration = 'penetration',
    EnergyRegeneration = 'energy-regeneration',
    AftershockDamage = 'aftershock-damage',
    DashAttackDamage = 'dash-attack-damage',
    ExSpecialAttackDamage = 'ex-special-attack-damage',
    AssistAttackDamage = 'assist-attack-damage',
    AnomalyBuildUpResistance = 'anomaly-build-up-resistance',
}